requirements.txt
asgiref==3.4.1
cffi==1.14.6
cryptography==35.0.0
Django==3.2.8
psycopg2-binary==2.9.1
pycparser==2.20
python-dotenv==0.19.0
pytz==2021.3
sqlparse==0.4.2
